+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # Version 1.0.2 (Feb 27, 2023)
 
 * fix libadapter binary searching in setup.py
+* fix setup.py for windows
 
 # Version 1.0.1 (Feb 26, 2023)